basic user permissions added

+++ posts/views.py
@@ -2,13 +2,14 @@ from urllib.parse import quote_plus
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib import messages
 from django.core.paginator import Paginator
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import Http404
 from .models import Post
 from .forms import PostForm
 
 # Create your views here.
 
 def post_view(request):
+    
     post_list = Post.objects.all()
     paginator = Paginator(post_list, 2) # Show 25 posts per page
     page_var = 'page'
@@ -18,6 +19,8 @@ def post_view(request):
 
 
 def post_create(request):
+    if not request.is_staff or not request.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     """
         create a instance of  PostFrom class from the forms module
@@ -42,6 +45,8 @@ def post_details(request, slug):
 
 
 def post_update(request, slug):
+    if not request.is_staff or not request.is_superuser:
+        raise Http404
     post = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=post)
     if form.is_valid and request.method == "POST":
@@ -60,6 +65,8 @@ def post_update(request, slug):
 
     
 def post_delete(request, slug):
+    if not request.is_staff or not request.is_superuser:
+        raise Http404
     post = get_object_or_404(Post, slug=slug)
     post.delete()
     messages.success(request, "Post deleted successfully")
